hardware_tracking/requirements.txt
Django==4.2.6
django-rest-swagger==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==19.12.0
flake8==6.1.0
pyflakes==3.1.0

pytest==7.4.3
pytest-django==4.5.2
python-dateutil==2.8.2

requests==2.31.0
responses==0.23.3
